update deeplink handling
